Settings and launch overhaul
This changes settings to handle launch separately from just calling a settings object. As so the launch can be done in a different thread. On launch, a logo has been added and there is a splashscreen of what folders it is searching through along with the version of nuke it has found. Currently it deletes the json and settings files associated with the program. This will need to be changed for release/packaging.