Fix early removal  issue, start on remove issue
The remove from list button does not work when the names are changed to just the filename. This is the start of an attempted workaround. It seems to somewhat work but there should be a check to make sure the dictionary doesn't get too big or remove some references if there remains items with the text.